Directory usage refactor
