+++ codegen/src/rendering.rb
@@ -1,0 +1,7 @@
+require 'active_support'
+require 'active_support/core_ext'
+require 'erb'
+
+def render_template(template, target, variables)
+
+end
