+++ codegen/templates/resources.rb.erb
@@ -1,0 +1,10 @@
+module Google
+  module Ads
+    module GoogleAds
+      module Factories
+        module Resource
+        end
+      end
+    end
+  end
+end
